refactor(python): helpers module changed to utils
