+++ src/python/src/database/env.py
@@ -7,7 +7,7 @@ from sqlalchemy import engine_from_config, pool
 import sys
 from pathlib import Path
 sys.path.append(str(Path(__file__).absolute().parent.parent))
-from helpers import mysql_connection_string  # noqa
+from utils import mysql_connection_string  # noqa
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
